Added lib dependency.
