--- app/src/main/java/com/krazijames/lib/view/adapter/MainActivity.java
@@ -1,13 +1,0 @@
-package com.krazijames.lib.view.adapter;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-    }
-}
