+++ example/src/main/java/com/krazijames/lib/view/adapter/example/MainActivity.java
@@ -1,0 +1,13 @@
+package com.krazijames.lib.view.adapter.example;
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+
+public class MainActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+    }
+}
