+++ example/src/main/res/values/dimens.xml
