+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':lib'
+include ':example', ':lib'
